--- backend/test_ran.py
@@ -1,5 +1,0 @@
-import os
-
-# get the current working directory
-working_dir = os.getcwd()
-print(working_dir)
